Update report_filters.def
Added schema necessary for filter view construction. Removed condition_column_name as this is not necessary in the BQ table, just helpful for 0-base counting for end users in the maintenance sheet. Changed condition_column_index to a STRING and added start_column_index as an INTEGER to support filter view formatting.